pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "discord-py>=2.5.2",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
]
